core: move bootstrap.ts to top-level subpath export
