+++ packages/core/src/bootstrap.ts
